Allow id to be given as alert options
This modifications allow to create alerts with a specific id. Also i slighly changed the remove method to be able to remove an alert with just a mocked object like `remove({id: xxx})`.

A use case to illustrate why this is needed (imo).
In my app i don't wanted to have `show()` call all other the code (especially in the websocket connector). So i decided to create a component that uses a redux state, this component is in charge of displaying and removing alerts based on what the state contains.
This very state is fed by a reducer reducer which subscribes to various event accross the whole application to listen for specific events.
Long story short, my component can't store currently shown alert's, nor my reducer. It means that once a notifications is shown i have no way to remove it (if there is no timeout).

This is the reason why i want to be able to control the id, because my reducer can generate id and store them to instruct the component to call `show` with this very id.